+++ drasyl-integration-tests/src/test/java/org/drasyl/peer/connection/server/ServerIT.java
@@ -26,7 +26,6 @@ import io.netty.util.ResourceLeakDetector;
 import io.reactivex.rxjava3.observers.TestObserver;
 import org.drasyl.DrasylConfig;
 import org.drasyl.DrasylException;
-import org.drasyl.DrasylNode;
 import org.drasyl.crypto.Crypto;
 import org.drasyl.crypto.CryptoException;
 import org.drasyl.identity.CompressedPrivateKey;
@@ -124,6 +123,9 @@ class ServerIT {
 
     @BeforeEach
     void setup(TestInfo info) throws DrasylException, CryptoException {
+//        ch.qos.logback.classic.Logger root = (ch.qos.logback.classic.Logger) LoggerFactory.getLogger("org.drasyl");
+//        root.setLevel(Level.TRACE);
+
         colorizedPrintln("STARTING " + info.getDisplayName(), COLOR_CYAN, STYLE_REVERSED);
 
         System.setProperty("io.netty.tryReflectionSetAccessible", "true");
@@ -133,7 +135,6 @@ class ServerIT {
         identitySession2 = Identity.of(26778671, "0236fde6a49564a0eaa2a7d6c8f73b97062d5feb36160398c08a5b73f646aa5fe5", "093d1ee70518508cac18eaf90d312f768c14d43de9bfd2618a2794d8df392da0");
 
         serverConfig = DrasylConfig.newBuilder()
-//                .loglevel(Level.TRACE)
                 .identityProofOfWork(ProofOfWork.of(6657650))
                 .identityPublicKey(CompressedPublicKey.of("023d34f317616c3bb0fa1e4b425e9419d1704ef57f6e53afe9790e00998134f5ff"))
                 .identityPrivateKey(CompressedPrivateKey.of("0c27af38c77f2cd5cc2a0ff5c461003a9c24beb955f316135d251ecaf4dda03f"))
@@ -148,7 +149,6 @@ class ServerIT {
                 .superPeerEnabled(false)
                 .messageMaxContentLength(1024 * 1024)
                 .build();
-        DrasylNode.setLogLevel(serverConfig.getLoglevel());
         serverIdentityManager = new IdentityManager(serverConfig);
         serverIdentityManager.loadOrCreateIdentity();
         peersManager = new PeersManager(event -> {
@@ -171,7 +171,6 @@ class ServerIT {
         server.open();
 
         configClient1 = DrasylConfig.newBuilder()
-//                .loglevel(Level.TRACE)
                 .serverEnabled(false)
                 .identityProofOfWork(identitySession1.getProofOfWork())
                 .identityPublicKey(identitySession1.getPublicKey())